+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: added AlwaysCheckRules method
 * 🎉 feat: added WhenNotExistOne + WhenNotExistAll
 
 1.3.5 (2023-08-19)